Move database module to creds module
